requirements.txt
astroid==2.1.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-cors-headers==2.4.0
django-import-export==1.2.0
djangorestframework==3.9.1
et-xmlfile==1.0.1
get==2018.11.19
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.5.14
post==2018.11.20
psycopg2==2.7.7
public==2018.11.20
pylint==2.2.2
pytz==2018.9
PyYAML==3.13
query-string==2018.11.20
request==2018.11.20
six==1.12.0
tablib==0.12.1
typed-ast==1.2.0
unicodecsv==0.14.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
